% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{one_of}
\alias{one_of}
\title{Superseded selection helpers}
\usage{
one_of(..., .vars = NULL)
}
\arguments{
\item{...}{One or more character vectors.}

\item{.vars}{A character vector of variable names. When called
from inside selecting functions like \code{\link[dplyr:select]{dplyr::select()}} these are
automatically set to the names of the table.}
}
\description{
\code{one_of()} is superseded in favour of the more precise \code{\link[=any_of]{any_of()}} and
\code{\link[=all_of]{all_of()}} selectors.
}
\keyword{internal}
