% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-rlang.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{quo}
\alias{quos}
\alias{enquo}
\alias{quo_name}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{rlang}{\code{\link[rlang]{enquo}}, \code{\link[rlang]{quo}}, \code{\link[rlang]{quo_name}}, \code{\link[rlang]{quos}}}
}}

