% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectr.R
\docType{package}
\name{selectr-package}
\alias{selectr}
\alias{selectr-package}
\title{selectr: Select From a Set of Strings}
\description{
The backend of selecting functions in the 'Tidyverse'
}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@rstudio.com}

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
