% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyselect.R
\docType{package}
\name{tidyselect-package}
\alias{tidyselect}
\alias{tidyselect-package}
\title{tidyselect: Select from a Set of Strings}
\description{
A backend for the selecting functions of the 'Tidyverse'
which makes it easy to implement select-like functions.
}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@rstudio.com}

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
