% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyselect.R
\docType{package}
\name{tidyselect-package}
\alias{tidyselect}
\alias{tidyselect-package}
\title{tidyselect: Select from a Set of Strings}
\description{
A backend for the selecting functions of the 'tidyverse'.
    It makes it easy to implement select-like functions in your own
    packages in a way that is consistent with other 'tidyverse'
    interfaces for selection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tidyselect.r-lib.org}
  \item \url{https://github.com/r-lib/tidyselect}
  \item Report bugs at \url{https://github.com/r-lib/tidyselect/issues}
}

}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@rstudio.com}

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
