% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{tidyselect_data_proxy}
\alias{tidyselect_data_proxy}
\alias{tidyselect_data_supports_predicates}
\title{tidyselect methods for custom types}
\usage{
tidyselect_data_proxy(x)

tidyselect_data_supports_predicates(x)
}
\arguments{
\item{x}{A data-frame like object passed to \code{\link[=eval_select]{eval_select()}},
\code{\link[=eval_rename]{eval_rename()}}, and friends.}
}
\description{
\itemize{
\item \code{tidyselect_data_proxy()} should return a 0-row data frame.
\item \code{tidyselect_data_supports_predicates()} should return \code{FALSE} if
\code{tidyselect_data_proxy()} doesn't return actual types.
}
}
