% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-deprecated.R
\name{vars_select}
\alias{vars_select}
\alias{vars_rename}
\title{Select or rename variables}
\usage{
vars_select(
  .vars,
  ...,
  .include = character(),
  .exclude = character(),
  .strict = TRUE
)

vars_rename(.vars, ..., .strict = TRUE)
}
\arguments{
\item{.vars}{A character vector of existing column names.}

\item{..., args}{Selection inputs. See the help for \link[=language]{selection helpers}.}

\item{.include, .exclude}{Character vector of column names to always
include/exclude.}

\item{.strict}{If \code{TRUE}, will throw an error if you attempt to select or
rename a variable that doesn't exist.}
}
\value{
A named character vector. Values are existing column names,
names are new names.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}

Please use \code{\link[=eval_select]{eval_select()}} and \code{\link[=eval_rename]{eval_rename()}} instead. See
\code{vignette("tidyselect")} to get started.
}
\seealso{
\code{\link[=vars_pull]{vars_pull()}}
}
\keyword{internal}
