% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{vars_select_helpers}
\alias{vars_select_helpers}
\title{List of selection helpers}
\format{
An object of class \code{list} of length 11.
}
\usage{
vars_select_helpers
}
\description{
This list contains all selection helpers exported in tidyselect. It
was useful when you wanted to embed the helpers in your API without
having to track addition of new helpers in tidyselect. However the
selection helpers are now always embedded in the DSL.
}
\keyword{internal}
